+++ src/app/components/Menu.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import "./styles/Menu.css";
+import "../styles/Menu.css";
 
 import {AppContext} from "./AppContext";
 import {useContext, useState} from "react";
