--- src/components/styles/Modal-ext.css
@@ -1,52 +1,0 @@
-
-
-.modal-container-ext {
-    position: fixed;
-    left: 0;
-    padding: 100px;
-    height: 100% ;
-    width: 100%;
-    background-color: rgba(0, 0, 0, 0.66);
-}
-
-.modal-ext {
-    position: fixed;
-    top: 15%;
-    left: 12%;
-
-    background-color: #fcdca9;
-    border-radius: 10px;
-    
-    height: 36rem;
-    width: 75%;
-}
-
-.modal-ext h2 {
-    position: fixed;
-    left: 17%;
-    top: 14%;
-    font-weight:900;
-}
-
-
-.close-modal-button-ext {
-    position: relative;
-    top: 2%;
-    left: 83%;
-
-
-    display: flex;
-    padding: 4px;
-    color: white;
-    text-align: center;
-    justify-content: center;
-    font-family:Arial;
-    font-size: 20px;
-    font-weight: bold
-    ;
-    background-color: rgb(225, 2, 2);
-    border-radius: 50%;
-    border: none;
-    height: 35px;
-    width: 35px;
-}
