--- src/components/styles/Modal.css
@@ -1,53 +1,0 @@
-
-
-.modal-container {
-    position: fixed;
-    padding: 200px;
-
-    height: 100% ;
-    width: 100%;
-    background-color: rgba(0, 0, 0, 0.66);
-}
-
-.modal {
-    position: fixed;
-    top: 18%;
-    left: 12%;
-
-    background-color: #fcdca9;
-    border-radius: 10px;
-    
-    height: 25rem;
-    width: 75%;
-}
-
-.modal h2 {
-    position: fixed;
-    left: 17%;
-    top: 19%;
-    font-weight:900;
-}
-
-.close-modal-button {
-    position: relative;
-    top: 5%;
-    left: 83%;
-
-
-    display: flex;
-    padding: 4px;
-    color: white;
-    text-align: center;
-    justify-content: center;
-    font-family:Arial;
-    font-size: 20px;
-    font-weight: bold
-    ;
-    background-color: rgb(225, 2, 2);
-    border-radius: 50%;
-    border: none;
-    height: 35px;
-    width: 35px;
-}
-
-
