--- src/components/styles/ModalQr.css
@@ -1,47 +1,0 @@
-
-
-
-.modal-container-qr {
-    position: fixed;
-    padding: 200px;
-    height: 100% ;
-    width: 100%;
-    background-color: rgba(0, 0, 0, 0.66);
-}
-
-.modal-qr {
-    position: fixed;
-    top: 10%;
-    left: 15%;
-    border-radius: 10px;
-    height: 20rem;
-    width: 71%;
-}
-
-.container-modal-qr {
-    margin-top: 20%;
-}
-
-.close-modal-button-qr {
-    position: relative;
-    top: 5%;
-    left: 90%;
-    display: flex;
-    padding: 4px;
-    color: white;
-    text-align: center;
-    justify-content: center;
-    font-family:Arial;
-    font-size: 20px;
-    font-weight: bold
-    ;
-    background-color: rgb(225, 2, 2);
-    border-radius: 50%;
-    border: none;
-    height: 35px;
-    width: 35px;
-}
-
-
-
-
