+++ src/controles/components/QrScan.jsx
