+++ src/controles/components/QrScan1.jsx
