+++ src/extintores/components/ExtintorCard.jsx
@@ -1,11 +1,10 @@
 import {useState} from 'react';
 
-import "./styles/Extintores.css"
+import "../styles/Extintores.css"
 
 const ExtintorCard = ({extintor, editExtintor, deleteExtintor, handleF_Vencimiento}) => {
 
   const [openExt, setOpenExt] = useState("ext-card");
-  const [content, setContent] = useState();
 
   const onTap = () => {
