+++ src/extintores/styles/Extintores.css
