+++ src/main.jsx
@@ -1,11 +1,11 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
-import NavBar from "./components/NavBar.jsx";
-import Login from './components/Login';
+import NavBar from "./app/components/NavBar.jsx";
+import Login from './app/components/Login';
 import { BrowserRouter } from "react-router-dom";
 
-import {AppProvider} from "./components/AppContext.jsx";
+import {AppProvider} from "./app/components/AppContext.jsx";
 
 createRoot(document.getElementById("root")).render(
 	// <StrictMode>
